--- Projects/Feature/HomeFeature/Sources/Sources.swift
@@ -1,1 +1,0 @@
-// Implementation module source code
